Auth/py-requirements.txt
sqlalchemy==1.3.11
rsa==4.0
PyJWT==1.5.3
psycopg2==2.8.4
pip==20.1
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
pyopenssl==19.1.0
passlib==1.7.2